some changes

+++ scripts/experiment_stmodel.sh
@@ -6,6 +6,7 @@
 # @brief    This script runs experiment for stmodel 
 #
 #   MODIFIED   (MM/DD/YY)
+#   stplaydog   08/12/16 - Also copy log for time analysis 
 #   stplaydog   08/06/16 - Parallel computing 
 #   stplaydog   04/27/16 - Add upload result script 
 #   stplaydog   04/24/16 - Add logger support 
@@ -16,10 +17,15 @@
 SCRIPT_HOME="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 OPTKIT_HOME=${SCRIPT_HOME}/../src/
 DATA_HOME=${SCRIPT_HOME}/../data/
+LOG_HOME=${SCRIPT_HOME}/../log/
 
 mkdir -p ${DATA_HOME}/experiments/stmodel/
+mkdir -p ${DATA_HOME}/experiments/stmodel/CHI
+mkdir -p ${DATA_HOME}/experiments/stmodel/DC
+mkdir -p ${DATA_HOME}/experiments/stmodel/NY
 
 rm ${SCRIPT_HOME}/process
+rm ${LOG_HOME}/*
 
 NUM_PROC=`nproc`
 
@@ -38,9 +44,13 @@ func_run_experiment()
                          -i ${DATA_HOME}stmodel/${WHICH_FILE}/${FILE}  \
                          -o ${DATA_HOME}experiments/stmodel/${WHICH_FILE}/${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME} \
                          -c ${XY} -c ${XY} -c ${TIME} \
-                         -l ${WHICH_FILE}_${XY}_${XY}_${TIME}_
+                         -l ${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME}_
 
-    cp ${DATA_HOME}experiments/stmodel/${WHICH_FILE}/${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME}.json ${DATA_HOME}experiments/results/
+    if [ -f ${DATA_HOME}experiments/stmodel/${WHICH_FILE}/${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME}.json ];
+    then
+        cp ${DATA_HOME}experiments/stmodel/${WHICH_FILE}/${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME}.json ${DATA_HOME}experiments/results/
+        cp ${LOG_HOME}/${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME}_* ${DATA_HOME}experiments/results/${WHICH_FILE}_${FILE}_${XY}_${XY}_${TIME}.log
+    fi
 
     echo END >> ${SCRIPT_HOME}/process
 }
@@ -48,7 +58,7 @@ func_run_experiment()
 #
 # For real crime test data
 #
-for WHICH_FILE in CHI DC NY
+for WHICH_FILE in DC NY CHI
 do
     for XY in 100 200 
     do
